+++ simply-synonyms/src/background/index.js
@@ -2,6 +2,7 @@ import firebase from 'firebase/app'
 import firebaseConfig from 'firebaseConfig'
 import setListeners from './setListeners'
 import initializeAuth from './auth'
+import browser from 'browserApi'
 
 firebase.initializeApp(firebaseConfig)